Fix email signup steps features test
This broke due to the disappearance of the catch all detailed
guide routes at the root. Now sending client to a functioning
url in Whitehall, as I do not know the url of the email sign up
page.

FIXME: replace mock_client url stub with email sign up page url