Add a way to track unpublishing of statistics announcements
At the moment, unpublishing a statistics announcement:

- adds a redirect to another announcement via Router API
- deletes the duplicate announcement completely

This means that Whitehall isn't the single source of info when it comes
to stats announcements and if the router lost its data somehow, it wouldn't
be possible to recreate the redirects.

This change introduces a DB column which would allow soft-deleting
announcements in Whitehall.